plugin/offers: hoist send_onion_reply into core.
We want to reuse it.
